requirements.txt
streamlit
pandas
plotly
openpyxl
python-docx
rapidfuzz
xlsxwriter
kaleido==0.2.1